refactor stories
